feat: Add Blog class
The Blog class represents a blog post and contains
information about its ID, title, the time and date
the post was published and if the blog post
contains an image, also the path to the image as
well as its alignment.
Moreover, this class a Category object with a
category ID and label and a User object with the
name, email address, location, ID and password.

This class works with getters and setters and contains
methods to fetch either all blog posts and blog posts
by category and to save a new blog post to the database.

It also contains a delegation to display the full name
of a user.