feat: Add Category class
This class represents a category and
contains information about its ID and the
name of the category.

This class works with getters and setters.

Moreover, it contains methods to check if a
category already exists in the database, to
fetch all categories from database and to
save a category to the database.