feat: Add User class
This class represents a user and
contains information about their name, ID,
email address, location and password for
their account on the website.

It works with getters and setters and has a
method to fetch the user data of a specific
user from the database.

Moreover, it contains a virtual attribute to
display the full name of a user.