feat: Add functionality to the edit form
The edit form can now be submitted and the blog
post will be updated in the database.

User error is made visible.

When images are deleted from the database while editing
or deleting a blog post, those images are deleted from
the server as well.

The user receives a confirmation about the result of the
editing.